% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-interrupt.R
\name{AsyncInterruptor}
\alias{AsyncInterruptor}
\title{An interruptor useful for stopping child processes.}
\description{
This class is a simple wrapper around a Queue object
making adding interrupt checking to future code easy
to implement and read.
}
\details{
\strong{Methods}
  \describe{
    \item{\code{initialize(queue=shinyQueue())}}{
      Creates a new interruptor.
    }
    \item{\code{interrupt(msg="Signaled Interrupt")}}{
      Signals an interrupt
    }
    \item{\code{execInterrupts()}}{
      Executes anything pushed to the queue, including interrupts.
    }
    \item{\code{getInterrupts()}}{
      Gets the result of the queue's executing, not throwing the interrupts.
    }
  }
}
\examples{
library(future)
strategy <- "future::multisession"
plan(strategy)
inter <- AsyncInterruptor$new()
fut <- future({
  for(i in 1:100){
    Sys.sleep(.01)
    inter$execInterrupts()
  }
})
inter$interrupt("Error: Stop Future")
try(value(fut))
inter$destroy()

# Clean up multisession cluster
plan(sequential)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AsyncInterruptor$new()}}
\item \href{#method-interrupt}{\code{AsyncInterruptor$interrupt()}}
\item \href{#method-execInterrupts}{\code{AsyncInterruptor$execInterrupts()}}
\item \href{#method-getInterrupts}{\code{AsyncInterruptor$getInterrupts()}}
\item \href{#method-destroy}{\code{AsyncInterruptor$destroy()}}
\item \href{#method-clone}{\code{AsyncInterruptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncInterruptor$new(queue = shinyQueue())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queue}}{a shiny queue}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-interrupt"></a>}}
\subsection{Method \code{interrupt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncInterruptor$interrupt(msg = "Signaled Interrupt")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{An error message string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execInterrupts"></a>}}
\subsection{Method \code{execInterrupts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncInterruptor$execInterrupts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInterrupts"></a>}}
\subsection{Method \code{getInterrupts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncInterruptor$getInterrupts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncInterruptor$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsyncInterruptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
