% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-progress.R
\docType{data}
\name{AsyncProgress}
\alias{AsyncProgress}
\title{A progress bar object where inc and set are usable within other processes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AsyncProgress
}
\description{
A progress bar object where inc and set are usable within other processes
}
\keyword{datasets}
