% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.R
\docType{data}
\name{Consumer}
\alias{Consumer}
\title{A Class for reading and executing tasks from a source}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Consumer
}
\description{
A Class for reading and executing tasks from a source
}
\keyword{datasets}
