% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{defaultSource}
\alias{defaultSource}
\title{Get/set the class used to sink/read from the file system}
\usage{
defaultSource(sourceClass)
}
\arguments{
\item{sourceClass}{An R6 object}
}
\description{
Get/set the class used to sink/read from the file system
}
