% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ipc-package}
\alias{ipc-package}
\title{Tools for performing async communication between workers in shiny}
\description{
Provides tools for passing messages between R processes. Shiny Examples are provided showing how to perform useful tasks such as:
updating reactive values from within a future, progress bars for long running async tasks, and interrupting async tasks based on user input.
}
\author{
Ian Fellows \email{ian@fellstat.com}
}
