% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{Queue}
\alias{Queue}
\alias{queue}
\title{A Class containing a producer and consumer}
\usage{
queue(
  source = defaultSource()$new(),
  producer = Producer$new(source),
  consumer = Consumer$new(source)
)
}
\arguments{
\item{source}{The source for reading and writing the queue}

\item{producer}{The producer for the source}

\item{consumer}{The consumer of the source}
}
\description{
Creates a Queue object for inter-process communication.
Its members \code{producer} and \code{consumer} are the main entry points for
sending and receiving messages respectively.
}
\details{
This function creates a queue object for communication between different R processes,
including forks of the same process.  It uses \code{txtq} backage as it's backend.
Technically, the information is sent using temporary files, created in a new directory
inside the session-specific temporary folder (see \code{\link{tempfile}}).
This requires that the new directory is writeable, this is normally the case but
if \code{\link{Sys.umask}} forbids writing, the communication fails with an error.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Queue$new()}}
\item \href{#method-destroy}{\code{Queue$destroy()}}
\item \href{#method-clone}{\code{Queue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$new(source, prod, cons)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
