% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{redisIdGenerator}
\alias{redisIdGenerator}
\title{Get/set the location for temporary files}
\usage{
redisIdGenerator(generator)
}
\arguments{
\item{generator}{a function generating id strings}
}
\description{
Get/set the location for temporary files
}
