% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis-source.R
\name{RedisSource}
\alias{RedisSource}
\title{Reads and writes the queue to a redis db}
\description{
Reads and writes the queue to a redis db

Reads and writes the queue to a redis db
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RedisSource$new()}}
\item \href{#method-getRedisConnection}{\code{RedisSource$getRedisConnection()}}
\item \href{#method-pop}{\code{RedisSource$pop()}}
\item \href{#method-push}{\code{RedisSource$push()}}
\item \href{#method-destroy}{\code{RedisSource$destroy()}}
\item \href{#method-finalize}{\code{RedisSource$finalize()}}
\item \href{#method-clone}{\code{RedisSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedisSource$new(id = redisIdGenerator()(), config = redisConfig())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{An identifier to use for the queue}

\item{\code{config}}{A configuration list for redux::hiredis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRedisConnection"></a>}}
\subsection{Method \code{getRedisConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedisSource$getRedisConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop"></a>}}
\subsection{Method \code{pop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedisSource$pop(n = -1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of records to pop (-1 indicates all available).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedisSource$push(msg, obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{A string indicating the signal.}

\item{\code{obj}}{The object to associate with the signal.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedisSource$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedisSource$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedisSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
