% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.R
\name{ShinyConsumer}
\alias{ShinyConsumer}
\title{A Consumer class with common task handlers useful in Shiny apps}
\description{
In addtion to 'eval' and 'function' signals, ShinyConsumer object
process 'interrupt' and 'notify' signals for throwing errors and
displying Shiny notifictions.
}
\section{Super class}{
\code{\link[ipc:Consumer]{ipc::Consumer}} -> \code{ShinyConsumer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-initHandlers}{\code{ShinyConsumer$initHandlers()}}
\item \href{#method-clone}{\code{ShinyConsumer$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="addHandler">}\href{../../ipc/html/Consumer.html#method-addHandler}{\code{ipc::Consumer$addHandler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="clearHandlers">}\href{../../ipc/html/Consumer.html#method-clearHandlers}{\code{ipc::Consumer$clearHandlers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="consume">}\href{../../ipc/html/Consumer.html#method-consume}{\code{ipc::Consumer$consume()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="finalize">}\href{../../ipc/html/Consumer.html#method-finalize}{\code{ipc::Consumer$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="getSource">}\href{../../ipc/html/Consumer.html#method-getSource}{\code{ipc::Consumer$getSource()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="initialize">}\href{../../ipc/html/Consumer.html#method-initialize}{\code{ipc::Consumer$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="removeHandler">}\href{../../ipc/html/Consumer.html#method-removeHandler}{\code{ipc::Consumer$removeHandler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="setSource">}\href{../../ipc/html/Consumer.html#method-setSource}{\code{ipc::Consumer$setSource()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="start">}\href{../../ipc/html/Consumer.html#method-start}{\code{ipc::Consumer$start()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="stop">}\href{../../ipc/html/Consumer.html#method-stop}{\code{ipc::Consumer$stop()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initHandlers"></a>}}
\subsection{Method \code{initHandlers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyConsumer$initHandlers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyConsumer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
