% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{shinyExample}
\alias{shinyExample}
\title{Run Example Shiny Apps}
\usage{
shinyExample(application = c("progress", "changeReactive", "cancel"))
}
\arguments{
\item{application}{The example to run}
}
\description{
Run Example Shiny Apps
}
\details{
'progress' is an example application with a long running analysis that is cancelable and has a progress bar.
'changeReaction' is the old faithful example, but with the histogram colors changing over time.
'cancel' is an example with a cancelable long running process.
}
