% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.R
\docType{data}
\name{ShinyProducer}
\alias{ShinyProducer}
\title{A Producer with methods specific for Shiny}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShinyProducer
}
\description{
A Producer with methods specific for Shiny
}
\keyword{datasets}
