% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{shinyQueue}
\alias{shinyQueue}
\title{Create a Queue object}
\usage{
shinyQueue(source = defaultSource()$new(),
  producer = ShinyProducer$new(source),
  consumer = ShinyConsumer$new(source))
}
\arguments{
\item{source}{The source for reading and writing the queue}

\item{producer}{The producer for the source}

\item{consumer}{The consumer of the source}
}
\description{
Create a Queue object
}
