% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file-source.R
\docType{data}
\name{ShinyTextFileSource}
\alias{ShinyTextFileSource}
\title{Reads and writes to a text file and destroys it on session end}
\description{
A wrapper around \code{txtq}.
}
\keyword{datasets}
