% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-interrupt.R
\name{stopMulticoreFuture}
\alias{stopMulticoreFuture}
\title{Stops a future run in a multicore plan}
\usage{
stopMulticoreFuture(x)
}
\arguments{
\item{x}{The MulticoreFuture}
}
\description{
Stops a future run in a multicore plan
}
\details{
This function sends terminate and kill signals to the process running the future,
and will only work for futures run on a multicore plan. This approach is not
recommended for cases where you can listen for interrupts within the future
(with \code{AsyncInterruptor}). However, for cases where long running code is
in an external library for which you don't have control, this can be the only way
to terminate the execution.
}
