% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{tempFileGenerator}
\alias{tempFileGenerator}
\title{Get/set the location for temporary files}
\usage{
tempFileGenerator(tempfile)
}
\arguments{
\item{tempfile}{a function generating working file path (e.g. tempfile())}
}
\description{
Get/set the location for temporary files
}
