% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ipc-package}
\alias{ipc-package}
\title{Tools for performing async communication between workers in shiny}
\description{
Tools for performing async communication between workers in shiny
}
\author{
Ian Fellows \email{ian@fellstat.com}
}
