% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.R
\docType{data}
\name{Producer}
\alias{Producer}
\title{A Class for sending tasks to a source}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Producer
}
\description{
A Class for sending tasks to a source
}
\keyword{datasets}
