% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{data}
\name{Queue}
\alias{Queue}
\title{A Class containing a producer and consumer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Queue
}
\description{
A Class containing a producer and consumer
}
\keyword{datasets}
