% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue}
\alias{queue}
\alias{Queue}
\title{Create a Queue object}
\usage{
queue(source = defaultSource()$new(), producer = Producer$new(source),
  consumer = Consumer$new(source))
}
\arguments{
\item{source}{The source for reading and writing the queue}

\item{producer}{The producer for the source}

\item{consumer}{The consumer of the source}
}
\description{
Create a Queue object
}
