% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis-source.R
\docType{data}
\name{RedisSource}
\alias{RedisSource}
\title{Reads and writes the queue to a redis db}
\arguments{
\item{id}{An identifier to use for the queue}

\item{config}{A configuration list for redux::hiredis}

\item{n}{The number of records to pop (-1 indicates all available).}

\item{msg}{A string indicating the signal.}

\item{obj}{The object to associate with the signal.}
}
\description{
Reads and writes the queue to a redis db
}
\keyword{datasets}
