% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.R
\docType{data}
\name{ShinyConsumer}
\alias{ShinyConsumer}
\title{A Consumer class with common task executors useful in Shiny apps}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShinyConsumer
}
\description{
A Consumer class with common task executors useful in Shiny apps
}
\keyword{datasets}
