% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.R
\docType{data}
\name{ShinyConsumer}
\alias{ShinyConsumer}
\title{A Consumer class with common task handlers useful in Shiny apps}
\description{
In addtion to 'eval' and 'function' signals, ShinyConsumer object
process 'interrupt' and 'notify' signals for throwing errors and
displying Shiny notifictions.
}
\keyword{datasets}
