% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.R
\name{ShinyConsumer}
\alias{ShinyConsumer}
\title{A Consumer class with common task handlers useful in Shiny apps}
\description{
A Consumer class with common task handlers useful in Shiny apps

A Consumer class with common task handlers useful in Shiny apps
}
\details{
In addtion to 'eval' and 'function' signals, ShinyConsumer object
process 'interrupt' and 'notify' signals for throwing errors and
displying Shiny notifictions.
}
\section{Super class}{
\code{\link[ipc:Consumer]{ipc::Consumer}} -> \code{ShinyConsumer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShinyConsumer-initHandlers}{\code{ShinyConsumer$initHandlers()}}
\item \href{#method-ShinyConsumer-clone}{\code{ShinyConsumer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="addHandler"><a href='../../ipc/html/Consumer.html#method-Consumer-addHandler'><code>ipc::Consumer$addHandler()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="clearHandlers"><a href='../../ipc/html/Consumer.html#method-Consumer-clearHandlers'><code>ipc::Consumer$clearHandlers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="consume"><a href='../../ipc/html/Consumer.html#method-Consumer-consume'><code>ipc::Consumer$consume()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="finalize"><a href='../../ipc/html/Consumer.html#method-Consumer-finalize'><code>ipc::Consumer$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="getSource"><a href='../../ipc/html/Consumer.html#method-Consumer-getSource'><code>ipc::Consumer$getSource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="initialize"><a href='../../ipc/html/Consumer.html#method-Consumer-initialize'><code>ipc::Consumer$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="removeHandler"><a href='../../ipc/html/Consumer.html#method-Consumer-removeHandler'><code>ipc::Consumer$removeHandler()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="setSource"><a href='../../ipc/html/Consumer.html#method-Consumer-setSource'><code>ipc::Consumer$setSource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="start"><a href='../../ipc/html/Consumer.html#method-Consumer-start'><code>ipc::Consumer$start()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Consumer" data-id="stop"><a href='../../ipc/html/Consumer.html#method-Consumer-stop'><code>ipc::Consumer$stop()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyConsumer-initHandlers"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyConsumer-initHandlers}{}}}
\subsection{Method \code{initHandlers()}}{
Adds default handlers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyConsumer$initHandlers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyConsumer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyConsumer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyConsumer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
