% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.R
\name{ShinyProducer}
\alias{ShinyProducer}
\title{A Producer with methods specific for Shiny}
\description{
A Producer object with additional methods for firing interrupts, shiny notifications,
and reactive value assignments.
}
\details{
\strong{Methods}
  \describe{
    \item{\code{fireInterrupt(msg="Interrupt")}}{
      Sends an error with message \code{msg}.
    }
    \item{\code{fireNotify(msg="Interrupt")}}{
      Sends a signal to create a shiy Notifiction with message \code{msg}.
    }
    \item{\code{fireAssignReactive(name, value)}}{
      Signals for assignment for reactive \code{name}  to \code{value}.
    }

    @param msg A string
    @param name The name of the reactive value.
    @param value The value to assign the reactive to.
 }
}
\section{Super class}{
\code{\link[ipc:Producer]{ipc::Producer}} -> \code{ShinyProducer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fireInterrupt}{\code{ShinyProducer$fireInterrupt()}}
\item \href{#method-fireNotify}{\code{ShinyProducer$fireNotify()}}
\item \href{#method-fireAssignReactive}{\code{ShinyProducer$fireAssignReactive()}}
\item \href{#method-clone}{\code{ShinyProducer$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fire">}\href{../../ipc/html/Producer.html#method-fire}{\code{ipc::Producer$fire()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fireCall">}\href{../../ipc/html/Producer.html#method-fireCall}{\code{ipc::Producer$fireCall()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fireDoCall">}\href{../../ipc/html/Producer.html#method-fireDoCall}{\code{ipc::Producer$fireDoCall()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fireEval">}\href{../../ipc/html/Producer.html#method-fireEval}{\code{ipc::Producer$fireEval()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="getSource">}\href{../../ipc/html/Producer.html#method-getSource}{\code{ipc::Producer$getSource()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="initialize">}\href{../../ipc/html/Producer.html#method-initialize}{\code{ipc::Producer$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="setSource">}\href{../../ipc/html/Producer.html#method-setSource}{\code{ipc::Producer$setSource()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fireInterrupt"></a>}}
\subsection{Method \code{fireInterrupt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$fireInterrupt(msg = "Interrupt")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fireNotify"></a>}}
\subsection{Method \code{fireNotify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$fireNotify(msg = "Notification")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fireAssignReactive"></a>}}
\subsection{Method \code{fireAssignReactive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$fireAssignReactive(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
