% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.R
\name{ShinyProducer}
\alias{ShinyProducer}
\title{A Producer with methods specific for Shiny}
\description{
A Producer with methods specific for Shiny

A Producer with methods specific for Shiny
}
\details{
A Producer object with additional methods for firing interrupts, shiny notifications,
and reactive value assignments.
}
\section{Super class}{
\code{\link[ipc:Producer]{ipc::Producer}} -> \code{ShinyProducer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShinyProducer-fireInterrupt}{\code{ShinyProducer$fireInterrupt()}}
\item \href{#method-ShinyProducer-fireNotify}{\code{ShinyProducer$fireNotify()}}
\item \href{#method-ShinyProducer-fireAssignReactive}{\code{ShinyProducer$fireAssignReactive()}}
\item \href{#method-ShinyProducer-clone}{\code{ShinyProducer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fire"><a href='../../ipc/html/Producer.html#method-Producer-fire'><code>ipc::Producer$fire()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fireCall"><a href='../../ipc/html/Producer.html#method-Producer-fireCall'><code>ipc::Producer$fireCall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fireDoCall"><a href='../../ipc/html/Producer.html#method-Producer-fireDoCall'><code>ipc::Producer$fireDoCall()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="fireEval"><a href='../../ipc/html/Producer.html#method-Producer-fireEval'><code>ipc::Producer$fireEval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="getSource"><a href='../../ipc/html/Producer.html#method-Producer-getSource'><code>ipc::Producer$getSource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="initialize"><a href='../../ipc/html/Producer.html#method-Producer-initialize'><code>ipc::Producer$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ipc" data-topic="Producer" data-id="setSource"><a href='../../ipc/html/Producer.html#method-Producer-setSource'><code>ipc::Producer$setSource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyProducer-fireInterrupt"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyProducer-fireInterrupt}{}}}
\subsection{Method \code{fireInterrupt()}}{
Sends an error with message \code{msg}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$fireInterrupt(msg = "Interrupt")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{A string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyProducer-fireNotify"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyProducer-fireNotify}{}}}
\subsection{Method \code{fireNotify()}}{
Sends a signal to create a shiny Notification with message \code{msg}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$fireNotify(msg = "Notification")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{A string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyProducer-fireAssignReactive"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyProducer-fireAssignReactive}{}}}
\subsection{Method \code{fireAssignReactive()}}{
Signals for assignment for reactive \code{name}  to \code{value}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$fireAssignReactive(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the reactive value.}

\item{\code{value}}{The value to assign the reactive to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyProducer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyProducer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyProducer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
