% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file-source.R
\docType{data}
\name{ShinyTextFileSource}
\alias{ShinyTextFileSource}
\title{Reads and writes the queue to a text file and destroys it on session end}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShinyTextFileSource
}
\description{
Reads and writes the queue to a text file and destroys it on session end
}
\details{
A thin wrapper around \code{txtq}
}
\keyword{datasets}
