% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file-source.R
\docType{data}
\name{TextFileSource}
\alias{TextFileSource}
\title{Reads and writes the queue to a text file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TextFileSource
}
\description{
Reads and writes the queue to a text file
}
\details{
A thin wrapper around \code{txtq}
}
\keyword{datasets}
