% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file-source.R
\docType{data}
\name{TextFileSource}
\alias{TextFileSource}
\title{Reads and writes the queue to a text file}
\arguments{
\item{filePath}{The path to the file}

\item{n}{The number of records to pop (-1 indicates all available).}

\item{msg}{A string indicating the signal.}

\item{obj}{The object to associate with the signal.}
}
\description{
A wrapper around \code{txtq}. This object saves signals
and associated objects to and queue, and retrieves them
for processing.
}
\keyword{datasets}
