% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file-source.R
\name{TextFileSource}
\alias{TextFileSource}
\title{Reads and writes the queue to a text file}
\description{
Reads and writes the queue to a text file

Reads and writes the queue to a text file
}
\details{
A wrapper around \code{txtq}. This object saves signals
and associated objects to and queue, and retrieves them
for processing.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TextFileSource-new}{\code{TextFileSource$new()}}
\item \href{#method-TextFileSource-pop}{\code{TextFileSource$pop()}}
\item \href{#method-TextFileSource-push}{\code{TextFileSource$push()}}
\item \href{#method-TextFileSource-destroy}{\code{TextFileSource$destroy()}}
\item \href{#method-TextFileSource-clone}{\code{TextFileSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFileSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-TextFileSource-new}{}}}
\subsection{Method \code{new()}}{
Creates a TextFileSource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFileSource$new(filePath = tempFileGenerator()())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filePath}}{The path to the file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFileSource-pop"></a>}}
\if{latex}{\out{\hypertarget{method-TextFileSource-pop}{}}}
\subsection{Method \code{pop()}}{
removes n items from the source and returns them
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFileSource$pop(n = -1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of records to pop (-1 indicates all available).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFileSource-push"></a>}}
\if{latex}{\out{\hypertarget{method-TextFileSource-push}{}}}
\subsection{Method \code{push()}}{
Adds an item to the source.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFileSource$push(msg, obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{A string indicating the signal.}

\item{\code{obj}}{The object to associate with the signal.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFileSource-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-TextFileSource-destroy}{}}}
\subsection{Method \code{destroy()}}{
Cleans up source after use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFileSource$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextFileSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TextFileSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFileSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
